# chore: faster sentry + use native polyfills